+++ components/o-teaser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [9.1.1](https://github.com/Financial-Times/origami/compare/o-teaser-v9.1.0...o-teaser-v9.1.1) (2025-11-18)
+
+
+### Bug Fixes
+
+* revert o3 focus styles on o-teaser links ([8d13d89](https://github.com/Financial-Times/origami/commit/8d13d89968893977a112b8c59779123fd43c9434))
+
 ## [9.1.0](https://github.com/Financial-Times/origami/compare/o-teaser-v9.0.0...o-teaser-v9.1.0) (2025-11-04)